requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
djangorestframework==3.14.0
greenlet==3.0.1
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
psycopg==3.1.12
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
